PLM PR1 review batch 2 fixes: comment cleanup, MIT headers, workspace deps, extract wxc audit module
- Drop future-PR/past-state phrasing from plm module doc-comments
- Prepend MIT copyright headers to coordination.rs, profile_gen.rs, wpr_path.rs
- Simplify plm log trace filename (drop PID and parse-preview note)
- Move serde_json/chrono/roxmltree to workspace [workspace.dependencies]
- Use wxc_common::string_util::from_wide in plm::wpr_path
- Add plm::coordination::singleton with shared try_acquire/release helpers;
  both plm::main and wxc::audit now delegate to it
- Extract PLM audit lifecycle (~230 lines) from wxc/src/main.rs into
  wxc/src/audit.rs
