+++ src/core/plm/Cargo.toml
@@ -17,15 +17,17 @@ anyhow.workspace = true
 # Portable deps (config / access_event / event_parser) must compile on
 # every target so their unit tests run in cross-platform CI. The
 # `windows` crate stays target-gated below.
-serde_json = { version = "1", features = ["preserve_order"] }
-chrono = { version = "0.4", default-features = false, features = ["std", "clock"] }
-roxmltree = "0.20"
+serde_json.workspace = true
+chrono.workspace = true
+roxmltree.workspace = true
 
 [target.'cfg(target_os = "windows")'.dependencies]
 windows = { workspace = true, features = [
     "Win32_System_EventLog",
     "Win32_System_SystemInformation",
+    "Win32_System_Threading",
 ] }
+wxc_common = { workspace = true }
 
 [build-dependencies]
 mxc_build_common.workspace = true
